Move cancel coverage behind a Tk-free seam
Extract batch loading and its cancel handle from the monolith/UI module so
the issue tests run without constructing Tk roots or dialogs.
