+++ src/hoi4cm/core/__init__.py
@@ -20,6 +20,7 @@ from hoi4cm.focus_tree import (
     ExportResult,
     ParsedFocusTree,
     apply_focus_code,
+    batch_load_trees,
     build_drawio_focuses,
     build_focus_name_lookup,
     build_focuses,
@@ -92,6 +93,7 @@ __all__ = [
     "append_scripted_loc",
     "autosave_path",
     "apply_focus_code",
+    "batch_load_trees",
     "bounded_inflate",
     "build_drawio_focuses",
     "build_focus_name_lookup",
